# docs/swagger.yaml
definitions:
...
  models.BookInventory:
    properties:
      authors:
        type: string
      available_copies:
        type: integer
      book_id:
        type: integer
      isbn:
        type: string
      lib_id:
        type: integer
      publisher:
        type: string
      title:
        type: string
      total_copies:
        type: integer
      version:
        type: string
    type: object
  models.Library:
    properties:
      id:
        type: integer
      name:
        type: string
    required:
    - name
    type: object
...
info:
  contact: {}
paths:
...
  /book:
    post:
      consumes:
      - application/json
      description: Create a new book in the library
      parameters:
      - description: Book data
        in: body
        name: book
        required: true
        schema:
          $ref: '#/definitions/models.BookInventory'
      produces:
      - application/json
      responses:
        "201":
          description: Book created successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid input
          schema:
            additionalProperties: true
            type: object
        "404":
          description: User not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to create book
          schema:
            additionalProperties: true
            type: object
      summary: Create a new book
      tags:
      - books
  /book/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a book by its ID
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Book deleted successfully
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Book not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to delete book
          schema:
            additionalProperties: true
            type: object
      summary: Delete a book by ID
      tags:
      - books
    get:
      consumes:
      - application/json
      description: Retrieve a book by its ID
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Book retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Book not found
          schema:
            additionalProperties: true
            type: object
      summary: Get a book by ID
      tags:
      - books
    put:
      consumes:
      - application/json
      description: Update a book's details by its ID
      parameters:
      - description: Book ID
        in: path
        name: id
        required: true
        type: string
      - description: Updated book data
        in: body
        name: book
        required: true
        schema:
          additionalProperties: true
          type: object
      produces:
      - application/json
      responses:
        "200":
          description: Book updated successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid input
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Book not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to update book
          schema:
            additionalProperties: true
            type: object
      summary: Update a book by ID
      tags:
      - books
  /books:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all books
      produces:
      - application/json
      responses:
        "200":
          description: Books retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to fetch books
          schema:
            additionalProperties: true
            type: object
      summary: Get all books
      tags:
      - books
  /issues:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all issues
      produces:
      - application/json
      responses:
        "200":
          description: All issues retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to fetch issues
          schema:
            additionalProperties: true
            type: object
      summary: Get all issues
      tags:
      - issues
  /library:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all libraries
      produces:
      - application/json
      responses:
        "200":
          description: Libraries fetched successfully
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to fetch libraries
          schema:
            additionalProperties: true
            type: object
      summary: Get all libraries
      tags:
      - library
    post:
      consumes:
      - application/json
      description: Create a new library and update the user's role to owner
      parameters:
      - description: Library data
        in: body
        name: library
        required: true
        schema:
          $ref: '#/definitions/models.Library'
      produces:
      - application/json
      responses:
        "201":
          description: Library created successfully
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Invalid input
          schema:
            additionalProperties: true
            type: object
        "404":
          description: User not found
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to create library
          schema:
            additionalProperties: true
            type: object
      summary: Create a new library
      tags:
      - library
...
  /return/{id}:
    put:
      consumes:
      - application/json
      description: Return a book by ISBN and reader ID from issue registry
      parameters:
      - description: Book ISBN
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Book returned successfully
          schema:
            additionalProperties: true
            type: object
        "401":
          description: User ID not found in context
          schema:
            additionalProperties: true
            type: object
        "404":
          description: Issue record not found
          schema:
            additionalProperties: true
            type: object
        "409":
          description: Book already returned
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to update issue record or book inventory
          schema:
            additionalProperties: true
            type: object
      summary: Return a book
      tags:
      - issues
  /search:
    get:
      consumes:
      - application/json
      description: Search books by title, author, or publisher
      parameters:
      - description: Book title
        in: query
        name: title
        type: string
      - description: Book authors
        in: query
        name: authors
        type: string
      - description: Book publisher
        in: query
        name: publisher
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Books retrieved successfully
          schema:
            additionalProperties: true
            type: object
        "500":
          description: Failed to fetch books
          schema:
            additionalProperties: true
            type: object
      summary: Search books
      tags:
      - books
...
swagger: "2.0"